Revert "Merge pull request #2162 from dgoodwin/query-gen-refactor"
This reverts commit 9cbf9b8d2768ebba65606bc960ee7465ce703468, reversing
changes made to 71ddbf9af91b0c446bb71240d3376b43e3f99da4.
