+++ pkg/api/componentreadiness/component_report.go
@@ -5,6 +5,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"regexp"
+	"slices"
 	"sort"
 	"strconv"
 	"strings"
@@ -27,6 +28,7 @@ import (
 	bqcachedclient "github.com/openshift/sippy/pkg/bigquery"
 	"github.com/openshift/sippy/pkg/componentreadiness/resolvedissues"
 	"github.com/openshift/sippy/pkg/regressionallowances"
+	"github.com/openshift/sippy/pkg/util/param"
 	"github.com/openshift/sippy/pkg/util/sets"
 )
 
@@ -40,6 +42,89 @@ const (
 	openRegressionConfidenceAdjustment = 5
 
 	explanationNoRegression = "No significant regressions found"
+
+	defaultJunitTable   = "junit"
+	rarelyRunJunitTable = "junit_rarely_run_jobs"
+
+	// This query de-dupes the test results. There are multiple issues present in
+	// our data set:
+	//
+	// 1. Some test suites in OpenShift retry, resulting in potentially multiple
+	//    failures for the same test in a job.  Component Readiness is currently
+	//    counting these as separate failures, resulting in an outsized impact on
+	//    our statistical analysis.
+	//
+	// 2. There is a second bug where successful test cases are sometimes
+	//    recorded by openshift-tests more than once, it's tracked by
+	//    https://issues.redhat.com/browse/OCPBUGS-16039
+	//
+	// 3. Flaked tests also have rows for the failures, so we need to ensure we only collect the flakes.
+	//
+	// 4. Flaked tests appear to be able to have success_val as 0 or 1.
+	//
+	// So, this sorts the data, partitioning by the 3-tuple of file_path/test_name/testsuite -
+	// preferring flakes, then successes, then failures, and we get the first row of each
+	// partition.
+	dedupedJunitTable = `
+		WITH deduped_testcases AS (
+			SELECT  
+				junit.*,
+				ROW_NUMBER() OVER(PARTITION BY file_path, test_name, testsuite ORDER BY
+					CASE
+						WHEN flake_count > 0 THEN 0
+						WHEN success_val > 0 THEN 1
+						ELSE 2
+					END) AS row_num,
+%s
+				jobs.org,
+				jobs.repo,
+				jobs.pr_number,
+				jobs.pr_sha,
+				CASE
+					WHEN flake_count > 0 THEN 0
+					ELSE success_val
+				END AS adjusted_success_val,
+				CASE
+					WHEN flake_count > 0 THEN 1
+					ELSE 0
+				END AS adjusted_flake_count
+			FROM
+				%s.%s AS junit
+			INNER JOIN %s.jobs  jobs ON 
+				junit.prowjob_build_id = jobs.prowjob_build_id 
+				AND jobs.prowjob_start >= DATETIME(@From)
+				AND jobs.prowjob_start < DATETIME(@To)
+			WHERE modified_time >= DATETIME(@From)
+			AND modified_time < DATETIME(@To)
+			AND skipped = false
+		)
+		SELECT * FROM deduped_testcases WHERE row_num = 1`
+
+	// normalJobNameCol simply uses the prow job name for regular (non-pull-request) component reports.
+	normalJobNameCol = `
+				jobs.prowjob_job_name AS variant_registry_job_name,
+`
+	// pullRequestDynamicJobNameCol is used for pull-request component reports and will use the releaseJobName
+	// annotation for /payload jobs if it exists, otherwise the normal prow job name. This is done as /payload
+	// jobs get custom job names which will not be in the variant registry.
+	pullRequestDynamicJobNameCol = `
+				CASE
+					WHEN EXISTS (
+						SELECT 1
+						FROM UNNEST(jobs.prowjob_annotations) AS annotation
+						WHERE annotation LIKE 'releaseJobName=%%'
+					) THEN (
+						SELECT
+						SPLIT(SPLIT(annotation, 'releaseJobName=')[OFFSET(1)], ',')[SAFE_OFFSET(0)]
+						FROM UNNEST(jobs.prowjob_annotations) AS annotation	
+						WHERE annotation LIKE 'releaseJobName=%%'
+						LIMIT 1
+					)
+					ELSE jobs.prowjob_job_name
+		    	END AS variant_registry_job_name,
+`
+	// consider fallback data good for 7 days
+	fallbackQueryTimeRoundingOverride = 24 * 7 * time.Hour
 )
 
 type GeneratorType string
@@ -300,11 +385,172 @@ func (c *componentReportGenerator) GenerateComponentReportTestStatus(ctx context
 	return componentReportTestStatus, nil
 }
 
+// sortedKeys is a helper that sorts the keys of a variant group map for consistent ordering.
+func sortedKeys[T any](it map[string]T) []string {
+	keys := make([]string, 0, len(it))
+	for k := range it {
+		keys = append(keys, k)
+	}
+	sort.StringSlice(keys).Sort()
+	return keys
+}
+
+// getCommonTestStatusQuery returns the common query for the higher level summary component summary.
+func (c *componentReportGenerator) getCommonTestStatusQuery(allJobVariants crtype.JobVariants, isSample, isFallback bool) (string, string, []bigquery.QueryParameter) {
+	// Parts of the query, including the columns returned, are dynamic, based on the list of variants we're told to work with.
+	// Variants will be returned as columns with names like: variant_[VariantName]
+	// See fetchTestStatus for where we dynamically handle these columns.
+	selectVariants := ""
+	joinVariants := ""
+	groupByVariants := ""
+	for _, v := range sortedKeys(allJobVariants.Variants) {
+		v = param.Cleanse(v) // should be clean already, but just in case
+		joinVariants += fmt.Sprintf("LEFT JOIN %s.job_variants jv_%s ON variant_registry_job_name = jv_%s.job_name AND jv_%s.variant_name = '%s'\n",
+			c.client.Dataset, v, v, v, v)
+	}
+	for _, v := range c.DBGroupBy.List() {
+		v = param.Cleanse(v)                                                        // should be clean already, but just in case
+		selectVariants += fmt.Sprintf("jv_%s.variant_value AS variant_%s,\n", v, v) // Note: Variants are camelcase, so the query columns come back like: variant_Architecture
+		groupByVariants += fmt.Sprintf("jv_%s.variant_value,\n", v)
+	}
+
+	jobNameQueryPortion := normalJobNameCol
+	if c.SampleRelease.PullRequestOptions != nil && isSample {
+		jobNameQueryPortion = pullRequestDynamicJobNameCol
+	}
+
+	// TODO: this is a temporary hack while we explore if rarely run jobs approach is actually going to work.
+	// A scheduled query is copying rarely run job results to a separate much smaller table every day, so we can
+	// query 3 months without spending a fortune. If this proves to work, we will work out a system of processing
+	// this as generically as we can, but it will be difficult.
+	junitTable := defaultJunitTable
+	for k, v := range c.IncludeVariants {
+		if k == "JobTier" {
+			if slices.Contains(v, "rare") {
+				junitTable = rarelyRunJunitTable
+			}
+		}
+	}
+
+	// WARNING: returning additional columns from this query will require explicit parsing in deserializeRowToTestStatus
+	// TODO: jira_component and jira_component_id appear to not be used? Could save bigquery costs if we remove them.
+	queryString := fmt.Sprintf(`WITH latest_component_mapping AS (
+						SELECT *
+						FROM %s.component_mapping cm
+						WHERE created_at = (
+								SELECT MAX(created_at)
+								FROM %s.component_mapping))
+					SELECT
+						ANY_VALUE(test_name) AS test_name,
+						ANY_VALUE(testsuite) AS test_suite,
+						cm.id as test_id,
+						%s
+						COUNT(cm.id) AS total_count,
+						SUM(adjusted_success_val) AS success_count,
+						SUM(adjusted_flake_count) AS flake_count,
+						MAX(CASE WHEN adjusted_success_val = 0 THEN modified_time ELSE NULL END) AS last_failure,
+						ANY_VALUE(cm.component) AS component,
+						ANY_VALUE(cm.capabilities) AS capabilities,
+					FROM (%s)
+					INNER JOIN latest_component_mapping cm ON testsuite = cm.suite AND test_name = cm.name
+`,
+		c.client.Dataset, c.client.Dataset, selectVariants, fmt.Sprintf(dedupedJunitTable, jobNameQueryPortion, c.client.Dataset, junitTable, c.client.Dataset))
+
+	queryString += joinVariants
+
+	groupString := fmt.Sprintf(`
+					GROUP BY
+						%s
+						cm.id `, groupByVariants)
+
+	queryString += `WHERE cm.staff_approved_obsolete = false AND 
+						(variant_registry_job_name LIKE 'periodic-%%' OR variant_registry_job_name LIKE 'release-%%' OR variant_registry_job_name LIKE 'aggregator-%%') 
+						AND NOT REGEXP_CONTAINS(variant_registry_job_name, @IgnoredJobs)`
+
+	commonParams := []bigquery.QueryParameter{
+		{
+			Name:  "IgnoredJobs",
+			Value: ignoredJobsRegexp,
+		},
+	}
+	if c.IgnoreDisruption {
+		queryString += ` AND NOT 'Disruption' in UNNEST(capabilities)`
+	}
+
+	// fallback queries get all variants with no filtering
+	// so all tests are fetched then cached
+	if !isFallback {
+		variantGroups := c.IncludeVariants
+		// potentially cross-compare variants for the sample
+		if isSample && len(c.VariantCrossCompare) > 0 {
+			variantGroups = c.CompareVariants
+		}
+		if variantGroups == nil { // server-side view definitions may omit a variants map
+			variantGroups = map[string][]string{}
+		}
+
+		for _, group := range sortedKeys(variantGroups) {
+			group = param.Cleanse(group) // should be clean already, but just to make sure
+			paramName := fmt.Sprintf("variantGroup_%s", group)
+			queryString += fmt.Sprintf(" AND (jv_%s.variant_value in UNNEST(@%s))", group, paramName)
+			commonParams = append(commonParams, bigquery.QueryParameter{
+				Name:  paramName,
+				Value: variantGroups[group],
+			})
+		}
+
+		for _, group := range sortedKeys(c.RequestedVariants) {
+			group = param.Cleanse(group) // should be clean already, but just to make sure
+			paramName := fmt.Sprintf("ReqVariant_%s", group)
+			queryString += fmt.Sprintf(` AND jv_%s.variant_value = @%s`, group, paramName)
+			commonParams = append(commonParams, bigquery.QueryParameter{
+				Name:  paramName,
+				Value: c.RequestedVariants[group],
+			})
+		}
+		if c.Capability != "" {
+			queryString += " AND @Capability in UNNEST(capabilities)"
+			commonParams = append(commonParams, bigquery.QueryParameter{
+				Name:  "Capability",
+				Value: c.Capability,
+			})
+		}
+		if c.TestID != "" {
+			queryString += ` AND cm.id = @TestId`
+			commonParams = append(commonParams, bigquery.QueryParameter{
+				Name:  "TestId",
+				Value: c.TestID,
+			})
+		}
+	}
+	return queryString, groupString, commonParams
+}
+
+type baseQueryGenerator struct {
+	client                   *bqcachedclient.Client
+	cacheOption              cache.RequestOptions
+	commonQuery              string
+	groupByQuery             string
+	queryParameters          []bigquery.QueryParameter
+	ComponentReportGenerator *componentReportGenerator
+}
+
 // getBaseQueryStatus builds the basis query, executes it, and returns the basis test status.
 func (c *componentReportGenerator) getBaseQueryStatus(ctx context.Context,
 	allJobVariants crtype.JobVariants) (map[string]crtype.TestStatus, []error) {
-
-	generator := newBaseQueryGenerator(c, allJobVariants)
+	baseQuery, baseGrouping, baseParams := c.getCommonTestStatusQuery(allJobVariants, false, false)
+	generator := baseQueryGenerator{
+		client: c.client,
+		cacheOption: cache.RequestOptions{
+			ForceRefresh: c.cacheOption.ForceRefresh,
+			// increase the time that base query is cached for since it shouldn't be changing?
+			CRTimeRoundingFactor: c.cacheOption.CRTimeRoundingFactor,
+		},
+		commonQuery:              baseQuery,
+		groupByQuery:             baseGrouping,
+		queryParameters:          baseParams,
+		ComponentReportGenerator: c,
+	}
 
 	componentReportTestStatus, errs := api.GetDataFromCacheOrGenerate[crtype.ReportTestStatus](ctx, c.client.Cache,
 		generator.cacheOption, api.GetPrefixedCacheKey("BaseTestStatus~", generator), generator.queryTestStatus, crtype.ReportTestStatus{})
@@ -316,16 +562,25 @@ func (c *componentReportGenerator) getBaseQueryStatus(ctx context.Context,
 	return componentReportTestStatus.BaseStatus, nil
 }
 
-func (c *componentReportGenerator) getFallbackBaseQueryStatus(ctx context.Context,
-	allJobVariants crtype.JobVariants,
-	release string, start, end time.Time) []error {
-	generator := newFallbackTestQueryReleasesGenerator(c, allJobVariants, release, start, end)
+func (c *componentReportGenerator) getFallbackBaseQueryStatus(ctx context.Context, allJobVariants crtype.JobVariants, release string, start, end time.Time) []error {
+	baseQuery, baseGrouping, baseParams := c.getCommonTestStatusQuery(allJobVariants, false, true)
+	generator := fallbackTestQueryReleasesGenerator{
+		client: c.client,
+		cacheOption: cache.RequestOptions{
+			ForceRefresh: c.cacheOption.ForceRefresh,
+			// increase the time that fallback queries are cached for
+			CRTimeRoundingFactor: fallbackQueryTimeRoundingOverride,
+		},
+		commonQuery:     baseQuery,
+		groupByQuery:    baseGrouping,
+		BaseRelease:     release,
+		BaseStart:       start,
+		BaseEnd:         end,
+		queryParameters: baseParams,
+		lock:            &sync.Mutex{},
+	}
 
-	cachedFallbackTestStatuses, errs := api.GetDataFromCacheOrGenerate[*crtype.FallbackReleases](
-		ctx, c.client.Cache, generator.cacheOption,
-		api.GetPrefixedCacheKey("FallbackReleases~", generator),
-		generator.getTestFallbackReleases,
-		&crtype.FallbackReleases{})
+	cachedFallbackTestStatuses, errs := api.GetDataFromCacheOrGenerate[*crtype.FallbackReleases](ctx, c.client.Cache, generator.cacheOption, api.GetPrefixedCacheKey("FallbackReleases~", generator), generator.getTestFallbackReleases, &crtype.FallbackReleases{})
 
 	if len(errs) > 0 {
 		return errs
@@ -335,10 +590,58 @@ func (c *componentReportGenerator) getFallbackBaseQueryStatus(ctx context.Contex
 	return nil
 }
 
+func (b *baseQueryGenerator) queryTestStatus(ctx context.Context) (crtype.ReportTestStatus, []error) {
+	before := time.Now()
+	errs := []error{}
+	baseString := b.commonQuery + ` AND branch = @BaseRelease`
+	baseQuery := b.client.BQ.Query(baseString + b.groupByQuery)
+
+	baseQuery.Parameters = append(baseQuery.Parameters, b.queryParameters...)
+	baseQuery.Parameters = append(baseQuery.Parameters, []bigquery.QueryParameter{
+		{
+			Name:  "From",
+			Value: b.ComponentReportGenerator.BaseRelease.Start,
+		},
+		{
+			Name:  "To",
+			Value: b.ComponentReportGenerator.BaseRelease.End,
+		},
+		{
+			Name:  "BaseRelease",
+			Value: b.ComponentReportGenerator.BaseRelease.Release,
+		},
+	}...)
+
+	baseStatus, baseErrs := fetchTestStatus(ctx, baseQuery)
+
+	if len(baseErrs) != 0 {
+		errs = append(errs, baseErrs...)
+	}
+
+	log.Infof("Base QueryTestStatus completed in %s with %d base results from db", time.Since(before), len(baseStatus))
+
+	return crtype.ReportTestStatus{BaseStatus: baseStatus}, errs
+}
+
+type sampleQueryGenerator struct {
+	client                   *bqcachedclient.Client
+	commonQuery              string
+	groupByQuery             string
+	queryParameters          []bigquery.QueryParameter
+	ComponentReportGenerator *componentReportGenerator
+}
+
 // getSampleQueryStatus builds the sample query, executes it, and returns the sample test status.
 func (c *componentReportGenerator) getSampleQueryStatus(
 	ctx context.Context, allJobVariants crtype.JobVariants) (map[string]crtype.TestStatus, []error) {
-	generator := newSampleQueryGenerator(c, allJobVariants)
+	commonQuery, groupByQuery, queryParameters := c.getCommonTestStatusQuery(allJobVariants, true, false)
+	generator := sampleQueryGenerator{
+		client:                   c.client,
+		commonQuery:              commonQuery,
+		groupByQuery:             groupByQuery,
+		queryParameters:          queryParameters,
+		ComponentReportGenerator: c,
+	}
 
 	componentReportTestStatus, errs := api.GetDataFromCacheOrGenerate[crtype.ReportTestStatus](ctx,
 		c.client.Cache, c.cacheOption,
@@ -350,6 +653,235 @@ func (c *componentReportGenerator) getSampleQueryStatus(
 	}
 
 	return componentReportTestStatus.SampleStatus, nil
+}
+
+func (s *sampleQueryGenerator) queryTestStatus(ctx context.Context) (crtype.ReportTestStatus, []error) {
+	before := time.Now()
+	errs := []error{}
+	sampleString := s.commonQuery + ` AND branch = @SampleRelease`
+	if s.ComponentReportGenerator.SampleRelease.PullRequestOptions != nil {
+		sampleString += `  AND org = @Org AND repo = @Repo AND pr_number = @PRNumber`
+	}
+	sampleQuery := s.client.BQ.Query(sampleString + s.groupByQuery)
+	sampleQuery.Parameters = append(sampleQuery.Parameters, s.queryParameters...)
+	sampleQuery.Parameters = append(sampleQuery.Parameters, []bigquery.QueryParameter{
+		{
+			Name:  "From",
+			Value: s.ComponentReportGenerator.SampleRelease.Start,
+		},
+		{
+			Name:  "To",
+			Value: s.ComponentReportGenerator.SampleRelease.End,
+		},
+		{
+			Name:  "SampleRelease",
+			Value: s.ComponentReportGenerator.SampleRelease.Release,
+		},
+	}...)
+	if s.ComponentReportGenerator.SampleRelease.PullRequestOptions != nil {
+		sampleQuery.Parameters = append(sampleQuery.Parameters, []bigquery.QueryParameter{
+			{
+				Name:  "Org",
+				Value: s.ComponentReportGenerator.SampleRelease.PullRequestOptions.Org,
+			},
+			{
+				Name:  "Repo",
+				Value: s.ComponentReportGenerator.SampleRelease.PullRequestOptions.Repo,
+			},
+			{
+				Name:  "PRNumber",
+				Value: s.ComponentReportGenerator.SampleRelease.PullRequestOptions.PRNumber,
+			},
+		}...)
+	}
+
+	sampleStatus, sampleErrs := fetchTestStatus(ctx, sampleQuery)
+
+	if len(sampleErrs) != 0 {
+		errs = append(errs, sampleErrs...)
+	}
+
+	log.Infof("Sample QueryTestStatus completed in %s with %d sample results db", time.Since(before), len(sampleStatus))
+
+	return crtype.ReportTestStatus{SampleStatus: sampleStatus}, errs
+}
+
+type fallbackTestQueryReleasesGenerator struct {
+	client                     *bqcachedclient.Client
+	cacheOption                cache.RequestOptions
+	commonQuery                string
+	groupByQuery               string
+	queryParameters            []bigquery.QueryParameter
+	allJobVariants             crtype.JobVariants
+	BaseRelease                string
+	BaseStart                  time.Time
+	BaseEnd                    time.Time
+	CachedFallbackTestStatuses crtype.FallbackReleases
+	lock                       *sync.Mutex
+}
+
+func (f *fallbackTestQueryReleasesGenerator) getTestFallbackReleases(ctx context.Context) (*crtype.FallbackReleases, []error) {
+	wg := sync.WaitGroup{}
+	f.CachedFallbackTestStatuses = newFallbackReleases()
+	releases, errs := GetReleaseDatesFromBigQuery(ctx, f.client)
+
+	if errs != nil {
+		return nil, errs
+	}
+
+	// currently gets current base plus previous 3
+	// current base is just for testing but use could be
+	// extended to no longer require the base query
+	var selectedReleases []*crtype.Release
+	fallbackRelease := f.BaseRelease
+
+	// Get up to 3 fallback releases
+	for i := 0; i < 3; i++ {
+		var crRelease *crtype.Release
+
+		fallbackRelease, err := previousRelease(fallbackRelease)
+		if err != nil {
+			log.WithError(err).Errorf("Failure determining fallback release for %s", fallbackRelease)
+			break
+		}
+
+		for i := range releases {
+			if releases[i].Release == fallbackRelease {
+				crRelease = &releases[i]
+				break
+			}
+		}
+
+		if crRelease != nil {
+			selectedReleases = append(selectedReleases, crRelease)
+		}
+	}
+
+	for _, crRelease := range selectedReleases {
+
+		start := f.BaseStart
+		end := f.BaseEnd
+
+		// we want our base release validation to match the base release report dates
+		if crRelease.Release != f.BaseRelease && crRelease.End != nil && crRelease.Start != nil {
+			end = *crRelease.End
+			start = *crRelease.Start
+		}
+
+		wg.Add(1)
+		go func(queryRelease crtype.Release, queryStart, queryEnd time.Time) {
+			defer wg.Done()
+			select {
+			case <-ctx.Done():
+				log.Infof("Context canceled while fetching fallback base query status")
+				return
+			default:
+				stats, errs := f.getTestFallbackRelease(ctx, queryRelease.Release, queryStart, queryEnd)
+				if len(errs) > 0 {
+					log.Errorf("FallbackBaseQueryStatus for %s failed with: %v", queryRelease, errs)
+					return
+				}
+
+				f.updateTestStatuses(queryRelease, stats.BaseStatus)
+			}
+		}(*crRelease, start, end)
+	}
+	wg.Wait()
+
+	return &f.CachedFallbackTestStatuses, nil
+}
+
+func (f *fallbackTestQueryReleasesGenerator) updateTestStatuses(release crtype.Release, updateStatuses map[string]crtype.TestStatus) {
+
+	var testStatuses crtype.ReleaseTestMap
+	var ok bool
+	// since we  can be called for multiple releases and
+	// we update the map below we need to block concurrent map writes
+	f.lock.Lock()
+	defer f.lock.Unlock()
+	if testStatuses, ok = f.CachedFallbackTestStatuses.Releases[release.Release]; !ok {
+		testStatuses = crtype.ReleaseTestMap{Release: release, Tests: map[string]crtype.TestStatus{}}
+		f.CachedFallbackTestStatuses.Releases[release.Release] = testStatuses
+	}
+
+	for key, value := range updateStatuses {
+		testStatuses.Tests[key] = value
+	}
+}
+
+type fallbackTestQueryGenerator struct {
+	client          *bqcachedclient.Client
+	cacheOption     cache.RequestOptions
+	commonQuery     string
+	groupByQuery    string
+	queryParameters []bigquery.QueryParameter
+	allJobVariants  crtype.JobVariants
+	BaseRelease     string
+	BaseStart       time.Time
+	BaseEnd         time.Time
+}
+
+func (f *fallbackTestQueryReleasesGenerator) getTestFallbackRelease(ctx context.Context, release string, start, end time.Time) (crtype.ReportTestStatus, []error) {
+	generator := fallbackTestQueryGenerator{
+		client: f.client,
+		cacheOption: cache.RequestOptions{
+			ForceRefresh: f.cacheOption.ForceRefresh,
+			// increase the time that base query is cached for since it shouldn't be changing
+			CRTimeRoundingFactor: fallbackQueryTimeRoundingOverride,
+		},
+		commonQuery:     f.commonQuery,
+		groupByQuery:    f.groupByQuery,
+		BaseRelease:     release,
+		BaseStart:       start,
+		BaseEnd:         end,
+		queryParameters: f.queryParameters,
+	}
+
+	testStatuses, errs := api.GetDataFromCacheOrGenerate[crtype.ReportTestStatus](ctx, f.client.Cache, generator.cacheOption, api.GetPrefixedCacheKey("FallbackBaseTestStatus~", generator), generator.getTestFallbackRelease, crtype.ReportTestStatus{})
+
+	if len(errs) > 0 {
+		return crtype.ReportTestStatus{}, errs
+	}
+
+	return testStatuses, nil
+}
+
+func (f *fallbackTestQueryGenerator) getTestFallbackRelease(ctx context.Context) (crtype.ReportTestStatus, []error) {
+	return f.getFallbackBaseQueryStatus(ctx, f.BaseRelease, f.BaseStart, f.BaseEnd)
+}
+
+func (f *fallbackTestQueryGenerator) getFallbackBaseQueryStatus(ctx context.Context, release string, start, end time.Time) (crtype.ReportTestStatus, []error) {
+	before := time.Now()
+	log.Infof("Starting Fallback (%s) QueryTestStatus", release)
+	errs := []error{}
+	baseString := f.commonQuery + ` AND branch = @BaseRelease`
+	baseQuery := f.client.BQ.Query(baseString + f.groupByQuery)
+
+	baseQuery.Parameters = append(baseQuery.Parameters, f.queryParameters...)
+	baseQuery.Parameters = append(baseQuery.Parameters, []bigquery.QueryParameter{
+		{
+			Name:  "From",
+			Value: start,
+		},
+		{
+			Name:  "To",
+			Value: end,
+		},
+		{
+			Name:  "BaseRelease",
+			Value: release,
+		},
+	}...)
+
+	baseStatus, baseErrs := fetchTestStatus(ctx, baseQuery)
+
+	if len(baseErrs) != 0 {
+		errs = append(errs, baseErrs...)
+	}
+
+	log.Infof("Fallback (%s) QueryTestStatus completed in %s with %d base results from db", release, time.Since(before), len(baseStatus))
+
+	return crtype.ReportTestStatus{BaseStatus: baseStatus}, errs
 }
 
 func (c *componentReportGenerator) getTestStatusFromBigQuery(ctx context.Context) (crtype.ReportTestStatus, []error) {
@@ -400,7 +932,6 @@ func (c *componentReportGenerator) getTestStatusFromBigQuery(ctx context.Context
 		}
 
 	}()
-
 	wg.Wait()
 	if len(baseErrs) != 0 || len(sampleErrs) != 0 {
 		errs = append(errs, baseErrs...)
@@ -484,6 +1015,44 @@ func (c *componentReportGenerator) getRowColumnIdentifications(testIDStr string,
 	columns = append(columns, crtype.ColumnID(columnKeyBytes))
 
 	return rows, columns, nil
+}
+
+func fetchTestStatus(ctx context.Context, query *bigquery.Query) (map[string]crtype.TestStatus, []error) {
+	errs := []error{}
+	status := map[string]crtype.TestStatus{}
+	log.Infof("Fetching test status with:\n%s\nParameters:\n%+v\n", query.Q, query.Parameters)
+
+	it, err := query.Read(ctx)
+	if err != nil {
+		log.WithError(err).Error("error querying test status from bigquery")
+		errs = append(errs, err)
+		return status, errs
+	}
+
+	for {
+		var row []bigquery.Value
+
+		err := it.Next(&row)
+		if err == iterator.Done {
+			break
+		}
+		if err != nil {
+			log.WithError(err).Error("error parsing component from bigquery")
+			errs = append(errs, errors.Wrap(err, "error parsing prowjob from bigquery"))
+			continue
+		}
+
+		testIDStr, testStatus, err := deserializeRowToTestStatus(row, it.Schema)
+		if err != nil {
+			err2 := errors.Wrap(err, "error deserializing row from bigquery")
+			log.Error(err2.Error())
+			errs = append(errs, err2)
+			continue
+		}
+
+		status[testIDStr] = testStatus
+	}
+	return status, errs
 }
 
 // deserializeRowToTestStatus deserializes a single row into a testID string and matching status.
@@ -626,7 +1195,7 @@ func (c *componentReportGenerator) normalizeProwJobName(prowName string) string 
 	return name
 }
 
-func (c *componentReportGenerator) fetchJobRunTestStatusResults(ctx context.Context,
+func (c *componentReportGenerator) fetchJobRunTestStatus(ctx context.Context,
 	query *bigquery.Query) (map[string][]crtype.
 	JobRunTestStatusRow, []error) {
 	errs := []error{}
